Add tag and branch validation to the convert.sh script
All branches and tags that are present in both git and cvs are now
compared if the --validate flag is passed to the convert.sh script.

The .gitignore file is now stripped out of the generated git
repository.

The README.md file has been updated based on the latest conversion
changes, all manual comparison tests have been repeated, and details
of the automated testing have been added.